# api/openapi.yaml
openapi: 3.1.0
info:
  title: shop-inventory-app API
  version: 0.2.0
  description: |
    Multi-channel inventory back-office. A single stock pool is sold through
    several channels (store, Shopee, LINE OA).

    Stock is tracked as on_hand and reserved, where available = on_hand -
    reserved. An order reserves every line or none of them, inside one
    transaction that locks the balance rows, so two channels can never sell
    the same unit.

    Every endpoint under /api except /api/auth/login and /api/auth/logout
    needs the session cookie set by login. Owner-only endpoints answer 403 for
    staff.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - session: []
paths:
...
  /api/products:
    get:
      tags: [products]
      summary: Every product with its stock balance
      operationId: listProducts
      parameters:
        - name: q
          in: query
          description: Case-insensitive match on name or SKU.
          schema: { type: string }
      responses:
        "200":
          description: Products ordered by SKU.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [products]
      summary: Create a product (owner)
      description: A new product starts with a zero balance; use stock-in to add units.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Another product already uses this SKU (code conflict).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/products/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [products]
      summary: One product, its balance and the open orders holding it
      operationId: getProduct
      responses:
        "200":
          description: Product detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [products]
      summary: Change product fields (owner)
      description: Only the fields sent are changed. Deactivating stops new orders but keeps existing reservations.
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sku: { type: string, maxLength: 40 }
                name: { type: string, maxLength: 200 }
                price: { type: string, pattern: '^\d{1,10}(\.\d{1,2})?$' }
                low_stock_threshold: { type: integer, minimum: 0 }
                is_active: { type: boolean }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Another product already uses this SKU.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/products/{id}/stock-in:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [stock]
      summary: Receive units into stock
      operationId: stockIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qty]
              properties:
                qty: { type: integer, minimum: 1, maximum: 100000 }
                note: { type: string, maxLength: 500 }
      responses:
        "200":
          description: New balance. A STOCK_IN movement was recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Validation" }
  /api/products/{id}/adjustments:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [stock]
      summary: Correct on-hand stock with a reason (owner)
      description: |
        A negative change can only take units that are available. Units held
        by orders cannot be adjusted away; the response is 409 with how many
        could have been removed.
      operationId: adjustStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [qty_change, reason]
              properties:
                qty_change: { type: integer, minimum: -100000, maximum: 100000, not: { const: 0 } }
                reason: { type: string, enum: [count_correction, damaged, lost, other] }
                note: { type: string, maxLength: 500, description: Required when reason is other. }
      responses:
        "200":
          description: New balance. An ADJUST movement was recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Balance" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Not enough available stock (code insufficient_stock).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                held:
                  value:
                    error:
                      code: insufficient_stock
                      message: not enough available stock
                      details: { requested: 4, available: 3, reserved: 5 }
        "422": { $ref: "#/components/responses/Validation" }
  /api/movements:
    get:
      tags: [stock]
      summary: The stock ledger, newest first
      operationId: listMovements
      parameters:
        - name: product_id
          in: query
          schema: { type: integer, format: int64 }
        - name: type
          in: query
          schema: { $ref: "#/components/schemas/MovementType" }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of movements. Each row carries the balance right after it.
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Movement" }
                  total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/users:
    get:
      tags: [users]
      summary: Team members (owner)
      operationId: listUsers
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/TeamMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [users]
      summary: Add a team member (owner)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, role, password]
              properties:
                email: { type: string, format: email }
                name: { type: string, maxLength: 100 }
                role: { $ref: "#/components/schemas/Role" }
                password: { type: string, minLength: 8 }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Email already has an account (code conflict).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422": { $ref: "#/components/responses/Validation" }
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    Unauthorized:
      description: No valid session (code unauthorized).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The signed-in role may not do this (code forbidden).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: No such resource (code not_found).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Validation:
      description: The request body failed validation (code validation_failed).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    Role:
      type: string
      enum: [owner, staff]
    User:
      type: object
      required: [id, email, name, role]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        name: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    TeamMember:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [created_at]
          properties:
            created_at: { type: string, format: date-time }
    Money:
      type: string
      pattern: '^\d{1,10}\.\d{2}$'
      description: Baht as a decimal string, never a float.
      examples: ["600.00"]
    ProductInput:
      type: object
      required: [sku, name, price]
      properties:
        sku: { type: string, pattern: '^[A-Za-z0-9._-]{1,40}$' }
        name: { type: string, maxLength: 200 }
        price: { type: string, pattern: '^\d{1,10}(\.\d{1,2})?$' }
        low_stock_threshold: { type: integer, minimum: 0, default: 0 }
        is_active: { type: boolean, default: true }
    StockStatus:
      type: string
      enum: [in_stock, low, out_of_stock]
      description: out_of_stock when available <= 0, low when available <= low_stock_threshold.
    Product:
      type: object
      required: [id, sku, name, price, low_stock_threshold, is_active, on_hand, reserved, available, stock_status, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        sku: { type: string }
        name: { type: string }
        price: { $ref: "#/components/schemas/Money" }
        low_stock_threshold: { type: integer }
        is_active: { type: boolean }
        on_hand: { type: integer }
        reserved: { type: integer }
        available: { type: integer, description: on_hand - reserved }
        stock_status: { $ref: "#/components/schemas/StockStatus" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Hold:
      type: object
      required: [order_id, order_no, channel, status, qty, created_at]
      properties:
        order_id: { type: integer, format: int64 }
        order_no: { type: string }
        channel: { $ref: "#/components/schemas/Channel" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        qty: { type: integer }
        created_at: { type: string, format: date-time }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          required: [holds]
          properties:
            holds:
              type: array
              description: Open orders (reserved or packed) holding this product.
              items: { $ref: "#/components/schemas/Hold" }
    Balance:
      type: object
      required: [product_id, on_hand, reserved, available]
      properties:
        product_id: { type: integer, format: int64 }
        on_hand: { type: integer }
        reserved: { type: integer }
        available: { type: integer }
    MovementType:
      type: string
      enum: [STOCK_IN, ADJUST, RESERVE, RELEASE, SHIP, RETURN]
    Movement:
      type: object
      required: [id, product_id, sku, product_name, type, qty_change, reserved_change, on_hand_after, reserved_after, available_after, created_at]
      properties:
        id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        sku: { type: string }
        product_name: { type: string }
        type: { $ref: "#/components/schemas/MovementType" }
        qty_change: { type: integer, description: Change to on_hand. }
        reserved_change: { type: integer, description: Change to reserved. }
        on_hand_after: { type: integer }
        reserved_after: { type: integer }
        available_after: { type: integer }
        order_id: { type: [integer, "null"], format: int64 }
        order_no: { type: [string, "null"] }
        order_channel:
          oneOf:
            - $ref: "#/components/schemas/Channel"
            - type: "null"
        reason: { type: [string, "null"], description: Adjust reason code. }
        note: { type: [string, "null"] }
        created_by_name: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
    Channel:
      type: string
      enum: [store, shopee, line]
      default: store
    OrderStatus:
      type: string
      enum: [reserved, packed, shipped, canceled]
...
    ErrorResponse:
      type: object
      required: [error]
      description: The envelope every failing endpoint returns.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code. Clients branch on this, not on message.
              enum:
                - bad_request
                - validation_failed
                - not_found
                - internal_error
                - service_unavailable
                - unauthorized
                - forbidden
                - conflict
                - invalid_credentials
                - insufficient_stock
                - invalid_state
            message:
              type: string
              description: Human-readable explanation. Not intended for programmatic use.
            request_id:
              type: string
              description: Correlates this response with the server logs.
            fields:
              type: array
              description: Present only for validation_failed.
              items:
                $ref: "#/components/schemas/FieldError"
            details:
              description: Code-specific data, such as the short lines of insufficient_stock.
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string